+++ Battleship/Assets/Scripts/Game/Client.cs
@@ -5,16 +5,18 @@ using UnityEngine;
 
 namespace Game
 {
-    public enum GamePhase { Build, Ready, Wait, Shoot, Ended }
+    public enum GamePhase
+    { Build, Ready, Wait, Shoot, Ended }
 
-    public class GameClient : NetworkBehaviour
+    public class Client : NetworkBehaviour
     {
         [Header("Controlling Server")]
-        public GameServer Server;
+        [SerializeField]
+        private Server Server;
 
         [Header("Boards")]
-        public GameBoard SelfBoard;
-        public GameBoard EnemyBoard;
+        public Board SelfBoard;
+        public Board EnemyBoard;
 
         [Header("Client Info")]
         public ulong Id;
@@ -23,7 +25,7 @@ namespace Game
         public string UserName;
 
         [Header("UI")]
-        public TMP_Text Text_TurnDisplay;
+        private TMP_Text Text_TurnDisplay;
 
         public void Start()
         {
@@ -59,43 +61,50 @@ namespace Game
                 case GamePhase.Build:
                     SelfBoard.BuildPhase();
                     break;
+
                 case GamePhase.Wait:
                     break;
+
                 case GamePhase.Shoot:
                     EnemyBoard.ShootingPhase();
                     break;
+
                 case GamePhase.Ended:
                     break;
+
                 default:
                     break;
             }
         }
 
-        [ClientRpc]
-        private void ChangePhaseClientRpc(GamePhase phase, ClientRpcParams clientRpcParams = default)
+        /// <summary>
+        /// Send client info and register on server
+        /// </summary>
+        /// <param name="clientId"></param>
+        /// <param name="userName"></param>
+        [ServerRpc]
+        private void SendInfoAndStartServerRpc(ulong clientId, string userName)
         {
-            Phase = phase;
+            Id = clientId;
+            UserName = userName;
+            SelfBoard.GameClient = this;
 
-            switch (Phase)
-            {
-                case GamePhase.Ready:
-                    Text_TurnDisplay.text = $"Waiting on '{EnemyBoard.Text_UserName.text}' to finish";
-                    break;
-                case GamePhase.Wait:
-                    Text_TurnDisplay.text = $"'{EnemyBoard.Text_UserName.text}' is shooting";
-                    break;
-                case GamePhase.Shoot:
-                    Text_TurnDisplay.text = $"Your turn";
-                    break;
-            }
+            Server = FindObjectOfType<Server>();
+
+            Server.RegisterPlayer(this);
         }
 
         [ClientRpc]
-        private void BuildNextShipClientRpc(PlaceShipDTO dto, ClientRpcParams clientRpcParams = default)
+        public void SendOppenentNameClientRpc(string userName, ClientRpcParams clientRpcParams = default)
         {
-            SelfBoard.BuildNextShip(dto);
+            EnemyBoard.Text_UserName.text = userName;
         }
 
+        /// <summary>
+        /// Checks if position of ship is valid, then places for both server and respective client.
+        /// </summary>
+        /// <param name="dto"></param>
+        /// <param name="rpcParams"></param>
         [ServerRpc]
         public void TryPlaceShipServerRpc(PlaceShipDTO dto, ServerRpcParams rpcParams = default)
         {
@@ -124,6 +133,16 @@ namespace Game
             }
         }
 
+        [ClientRpc]
+        private void BuildNextShipClientRpc(PlaceShipDTO dto, ClientRpcParams clientRpcParams = default)
+        {
+            SelfBoard.BuildNextShip(dto);
+        }
+
+        /// <summary>
+        /// Checks for all ships being placed, then checks for 'Server GameStart'
+        /// </summary>
+        /// <returns></returns>
         private bool HasBuildPhaseEnded()
         {
             if (SelfBoard.Ships.Count is 5)
@@ -138,24 +157,10 @@ namespace Game
             return false;
         }
 
-        [ClientRpc]
-        public void SendOppenentNameClientRpc(string userName, ClientRpcParams clientRpcParams = default)
-        {
-            EnemyBoard.Text_UserName.text = userName;
-        }
-
-        [ServerRpc]
-        private void SendInfoAndStartServerRpc(ulong clientId, string userName)
-        {
-            Id = clientId;
-            UserName = userName;
-            SelfBoard.GameClient = this;
-
-            Server = FindObjectOfType<GameServer>();
-
-            Server.RegisterPlayer(this);
-        }
-
+        /// <summary>
+        /// Changes Phase on Server side and Client side
+        /// </summary>
+        /// <param name="phase"></param>
         public void ChangePhase(GamePhase phase)
         {
             Phase = phase;
@@ -169,15 +174,37 @@ namespace Game
             });
         }
 
-        [ServerRpc]
-        public void SendShotToServerRpc(Vector2 gridPosition)
+        /// <summary>
+        /// Change the current phase on the clients side
+        /// </summary>
+        /// <param name="phase"></param>
+        /// <param name="clientRpcParams"></param>
+        [ClientRpc]
+        private void ChangePhaseClientRpc(GamePhase phase, ClientRpcParams clientRpcParams = default)
         {
-            Server.SendAttack(gridPosition, this);
+            Phase = phase;
 
-            Destroy(EnemyBoard.MouseMarker);
-            EnemyBoard.MouseMarker = null;
+            switch (Phase)
+            {
+                case GamePhase.Ready:
+                    Text_TurnDisplay.text = $"Waiting on '{EnemyBoard.Text_UserName.text}' to finish";
+                    break;
+
+                case GamePhase.Wait:
+                    Text_TurnDisplay.text = $"'{EnemyBoard.Text_UserName.text}' is shooting";
+                    break;
+
+                case GamePhase.Shoot:
+                    Text_TurnDisplay.text = $"Your turn";
+                    break;
+            }
         }
 
+        /// <summary>
+        /// Checks through all ships and their respective positions for a hit
+        /// </summary>
+        /// <param name="gridPosition"></param>
+        /// <returns></returns>
         public bool AnyShipsHit(Vector2 gridPosition)
         {
             foreach (var shipObject in SelfBoard.Ships)
@@ -194,12 +221,24 @@ namespace Game
             return false;
         }
 
+        [ServerRpc]
+        public void SendShotToServerRpc(Vector2 gridPosition)
+        {
+            Server.SendAttack(gridPosition, this);
+        }
+
         [ClientRpc]
         public void InstantiateHitmarkerClientRpc(Vector2 gridPosition, bool isHit, bool isShooter)
         {
             var board = isShooter ? EnemyBoard : SelfBoard;
 
             board.InstantiateHitmarker(gridPosition, isHit);
+        }
+
+        [ClientRpc]
+        public void DisconnectClientRpc()
+        {
+            Text_TurnDisplay.text = $"Your opponent has disconnected - The game will reset";
         }
 
         [ClientRpc]
@@ -210,12 +249,6 @@ namespace Game
             Text_TurnDisplay.text = $"'{winningUser}' has won!";
 
             //TODO: Winning logic
-        }
-
-        [ClientRpc]
-        public void DisconnectClientRpc()
-        {
-            Text_TurnDisplay.text = $"Your opponent has disconnected - The game will reset";
         }
     }
 }
